fixed flutter map and supabase auth

+++ lib/data/services/supabase_service.dart
@@ -1,9 +1,20 @@
 // data/services/supabase_service.dart
+import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 import '../../core/constants/app_constants.dart';
+import '../models/user_model.dart';
 
 class SupabaseService {
   late final SupabaseClient _client;
+  late final SupabaseStorageClient _storage;
+final FlutterSecureStorage _secureStorage = const FlutterSecureStorage();
+  static final SupabaseService _instance = SupabaseService._internal();
+
+  factory SupabaseService() {
+    return _instance;
+  }
+
+  SupabaseService._internal();
 
   Future<void> initialize() async {
     _client = Supabase.instance.client;
@@ -20,12 +31,24 @@ class SupabaseService {
     final response = await _client.auth.signUp(
       email: email,
       password: password,
-      data: {
-        'full_name': fullName,
-        'role': AppConstants.roleClient,
-      },
+      
     );
+     final supabaseUser = _client.auth.currentUser!.id;
+      await _client.from('profiles').insert(
+        {
+      'user_id': supabaseUser,
+      'full_name': fullName,
+      'role': AppConstants.roleClient,
+      'created_at': DateTime.now().toIso8601String(),
+      'avatar_url': 'https://avatars.githubusercontent.com/u/1200?v=4',
+      'phone': '',                                       
+    },
+      );
+        if (response.user == null) {
+      print("State Massages : ${response.user}");
+    }
     return response;
+    
   }
 
   Future<AuthResponse> signIn({
@@ -36,8 +59,38 @@ class SupabaseService {
       email: email,
       password: password,
     );
+    final uidUser = _client.auth.currentUser!.id;
+    print("State Massages : ${response.user}");
+    if (response.user != null) {
+      print("State Massages : ${response.user}");
+      final res = await _client.from('profiles').select().match({'user_id': uidUser}).maybeSingle();
+      
+      if (res == null) {
+          print("User profile not found");
+          return response;
+      }
+      if (res.isEmpty) {
+        print("User profile data is empty");
+          return response;
+      }
+        final user = UserModel.fromJson(res);
+        await _secureStorage.write(key: 'user_id', value: user.id);
+       await _secureStorage.write(key: 'user_role', value: user.role);
+        print("User Role: ${user.role}");
+        print("User ID: ${user.id}");
+    }
     return response;
   }
+//   Future<String?> getUserRole() async {
+//   final user = _client.auth.currentUser;
+
+//   if (user == null) return null;
+
+  
+
+//   return res?['role'];
+// }
+
 
   Future<void> signOut() async {
     await _client.auth.signOut();
@@ -74,15 +127,15 @@ class SupabaseService {
     await _client
         .from(AppConstants.parkingSpotsTable)
         .update({'is_available': isAvailable})
-        .eq('id', id);
+        .eq('user_id', id);
   }
 
-  Future<Map<String, dynamic>> fetchUserProfile(String userId) async {
+  Future<Map<String, dynamic>?> fetchUserProfile(String userId) async {
     final response = await _client
         .from(AppConstants.usersTable)
         .select()
-        .eq('id', userId)
-        .single();
+        .eq('user_id', userId)
+        .maybeSingle();
     return response;
   }
 
@@ -90,7 +143,7 @@ class SupabaseService {
     await _client
         .from(AppConstants.usersTable)
         .update(userData)
-        .eq('id', userId);
+        .eq('user_id', userId);
   }
 
   // Admin methods
@@ -114,14 +167,14 @@ class SupabaseService {
     await _client
         .from(AppConstants.parkingSpotsTable)
         .update(parkingData)
-        .eq('id', id);
+        .eq('user_id', id);
   }
 
   Future<void> deleteParkingSpot(String id) async {
     await _client
         .from(AppConstants.parkingSpotsTable)
         .delete()
-        .eq('id', id);
+        .eq('user_id', id);
   }
 
   Future<Map<String, dynamic>> fetchDashboardStats() async {
